+++ src/main/java/tricatch/oe/proxy/pass/RelayWebSocket.java
@@ -28,7 +28,7 @@ public class RelayWebSocket {
      * @return HttpStream.Connection indicating whether connection should be closed
      * @throws IOException when I/O error occurs
      */
-    public static HttpStream.Connection relay(String clientId, String rid, HttpStream.Flow flow, HttpStreamReader in, HttpStreamWriter out) throws IOException {
+    public static HttpStream.Connection relay(String clientId, String rid, HttpStream.Flow flow, HttpStreamReader in, HttpStreamWriter out, boolean monitored) throws IOException {
         if (logger.isDebugEnabled()) {
             logger.debug("{}, {}, Relaying WebSocket frames"
                     , rid